backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0

# Authentication
firebase-admin>=6.4.0,<7.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Validation and settings
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# AI
anthropic>=0.18.0,<1.0.0

# HTTP client
httpx>=0.26.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pytz>=2024.1

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
aiosqlite>=0.19.0,<1.0.0

# Development
black>=24.1.0,<25.0.0
ruff>=0.2.0,<1.0.0
